mssql support

+++ _build-and-test-all-cdc.sh
@@ -30,63 +30,68 @@ $DOCKER_COMPOSE up -d cdcservice
 ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-in-memory:cleanTest :eventuate-tram-e2e-tests-in-memory:test
 ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-kafka:cleanTest :eventuate-tram-e2e-tests-jdbc-kafka:test
 
-$DOCKER_COMPOSE stop kafka
 
-if [[ "${DATABASE}" == "mysql" ]]; then
-    $DOCKER_COMPOSE up -d activemq
+if [[ "${DATABASE}" != "mssql" ]]; then
+
+    $DOCKER_COMPOSE stop kafka
+
+    if [[ "${DATABASE}" == "mysql" ]]; then
+        $DOCKER_COMPOSE up -d activemq
+        $DOCKER_COMPOSE stop cdcservice
+        $DOCKER_COMPOSE rm --force cdcservice
+
+        if [ -z "$SPRING_PROFILES_ACTIVE" ] ; then
+          export SPRING_PROFILES_ACTIVE=ActiveMQ
+        else
+          export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE},ActiveMQ
+        fi
+
+        $DOCKER_COMPOSE up -d cdcservice
+        ./wait-for-services.sh $DOCKER_HOST_IP 8099
+
+        ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-activemq:cleanTest :eventuate-tram-e2e-tests-jdbc-activemq:test
+
+        $DOCKER_COMPOSE stop activemq
+        $DOCKER_COMPOSE up -d rabbitmq
+        $DOCKER_COMPOSE stop cdcservice
+        $DOCKER_COMPOSE rm --force cdcservice
+
+        export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/ActiveMQ/RabbitMQ}
+
+        $DOCKER_COMPOSE up -d cdcservice
+        ./wait-for-services.sh $DOCKER_HOST_IP 8099
+
+        ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-rabbitmq:cleanTest :eventuate-tram-e2e-tests-jdbc-rabbitmq:test
+
+
+        $DOCKER_COMPOSE stop rabbitmq
+        export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/RabbitMQ/Redis}
+    else
+        if [ -z "$SPRING_PROFILES_ACTIVE" ] ; then
+          export SPRING_PROFILES_ACTIVE=Redis
+        else
+          export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE},Redis
+        fi
+    fi
+
+    $DOCKER_COMPOSE stop zookeeper
+    $DOCKER_COMPOSE up -d redis
     $DOCKER_COMPOSE stop cdcservice
     $DOCKER_COMPOSE rm --force cdcservice
 
-    if [ -z "$SPRING_PROFILES_ACTIVE" ] ; then
-      export SPRING_PROFILES_ACTIVE=ActiveMQ
-    else
-      export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE},ActiveMQ
-    fi
 
     $DOCKER_COMPOSE up -d cdcservice
     ./wait-for-services.sh $DOCKER_HOST_IP 8099
 
-    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-activemq:cleanTest :eventuate-tram-e2e-tests-jdbc-activemq:test
+    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-redis:cleanTest :eventuate-tram-e2e-tests-jdbc-redis:test
 
-    $DOCKER_COMPOSE stop activemq
-    $DOCKER_COMPOSE up -d rabbitmq
-    $DOCKER_COMPOSE stop cdcservice
-    $DOCKER_COMPOSE rm --force cdcservice
+    $DOCKER_COMPOSE stop redis
+    sleep 10
+    $DOCKER_COMPOSE start redis
 
-    export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/ActiveMQ/RabbitMQ}
+    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-redis:cleanTest :eventuate-tram-e2e-tests-jdbc-redis:test
 
-    $DOCKER_COMPOSE up -d cdcservice
-    ./wait-for-services.sh $DOCKER_HOST_IP 8099
-
-    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-rabbitmq:cleanTest :eventuate-tram-e2e-tests-jdbc-rabbitmq:test
-
-
-    $DOCKER_COMPOSE stop rabbitmq
-    export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/RabbitMQ/Redis}
-else
-    if [ -z "$SPRING_PROFILES_ACTIVE" ] ; then
-      export SPRING_PROFILES_ACTIVE=Redis
-    else
-      export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE},Redis
-    fi
 fi
-
-$DOCKER_COMPOSE stop zookeeper
-$DOCKER_COMPOSE up -d redis
-$DOCKER_COMPOSE stop cdcservice
-$DOCKER_COMPOSE rm --force cdcservice
-
-
-$DOCKER_COMPOSE up -d cdcservice
-./wait-for-services.sh $DOCKER_HOST_IP 8099
-
-./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-redis:cleanTest :eventuate-tram-e2e-tests-jdbc-redis:test
-
-$DOCKER_COMPOSE stop redis
-sleep 10
-$DOCKER_COMPOSE start redis
-
-./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-redis:cleanTest :eventuate-tram-e2e-tests-jdbc-redis:test
 
 $DOCKER_COMPOSE stop
 $DOCKER_COMPOSE rm --force -v
